requirements.txt
customtkinter==5.2.2
requests==2.32.3
Pillow==11.0.0
python-dotenv==1.0.1
cryptography==44.0.0
